refactor: dealer lose할 경우의 수 추가
